added custom hook to avoid repeating same logic
