+++ src/pages/pipeline/ingest.tsx
@@ -7,7 +7,7 @@ import { Sources } from '@/components/Sources/Sources';
 import { Destinations } from '@/components/Destinations/Destinations';
 import { PageHead } from '@/components/PageHead';
 import { ConnectionSyncLogsProvider } from '@/contexts/ConnectionSyncLogsContext';
-import { useRouter } from 'next/router';
+import { useQueryParams } from '@/customHooks/useQueryParams';
 
 interface TabPanelProps {
   children?: React.ReactNode;
@@ -32,33 +32,17 @@ function TabPanel(props: TabPanelProps) {
 }
 
 export default function Ingest() {
-  const router = useRouter();
-  const { tab }: any = router.query;
-
   const tabsObj: { [key: string]: number } = {
     connections: 0,
     sources: 1,
     warehouse: 2
   }
-  const reverseTabsObj = Object.entries(tabsObj).reduce(
-    (acc, [key, value]) => ({ ...acc, [value]: key }),
-    {} as { [key: number]: string }
-  );
-  const currentTab = tab ? tabsObj[tab] : 0;
-  const [value, setValue] = React.useState(currentTab);
 
-  useEffect(() => {
-    if (!tab || !tabsObj[tab]) {
-      router.push(`/pipeline/ingest?tab=${"connections"}`, undefined, { shallow: true });
-    }
-    setValue(currentTab);
-  }, [currentTab]);
-
-  const handleChange = (event: React.SyntheticEvent, newValue: number) => {
-    setValue(newValue);
-    const tabName = reverseTabsObj[newValue];
-    router.push(`/pipeline/ingest?tab=${tabName}`, undefined, { shallow: true });
-  };
+  const { value, handleChange } = useQueryParams({
+    tabsObj,
+    basePath: "/pipeline/ingest",
+    defaultTab: "connections"
+  })
 
   return (
     <>
